Added Truck and Lorry classes as obstacle vehicles, updated movement so it's smoother. Started using a json file to control the progress of the game and the spawns of vehicles

+++ lib/roadfighter/Entity.h
@@ -24,9 +24,9 @@ namespace roadfighter {
 
         virtual void draw() {};
 
-        virtual int getSpeed() const {};
+        virtual double getSpeed() const {};
 
-        virtual void setSpeed(int speed) {};
+        virtual void setSpeed(double speed) {};
 
         virtual int getMax_speed() const {};
     };
